Add notes on upcoming deprecations to release notes
The 5.0 release, which is coming up fast, will make a number of
deprecations and removals. We can start giving folks a heads up
on them now, so there is no confusion when 5.0 drops. We will
also write some blogs about the transition over the coming weeks
(after we have a better idea of what the final release will look
like).

Signed-off-by: Matt Heon <mheon@redhat.com>