Set stdin to nil for HyperV execs
We're seeing a bug where failed Podman commands can hang in the
test suite, but somehow un-hang on STDIN activity (e.g. pressing
enter). On Brent's vague suspicion that it was related to STDIN,
I disabled STDIN on all execs from HyperV, and the hangs
disappear. I don't know why this is, or why the hangs happen in
the first place, but since this seems to work and should not have
any negative consequences that I can think of, here we go.

[NO NEW TESTS NEEDED] I have no idea how to test this one.

Signed-off-by: Matt Heon <mheon@redhat.com>