machine: Save machine serial console by default
Motivated by making it easier to debug problems like
the applehv issues we're seeing today.  This is also
easy to do with qemu.

However note that in practice this really
wants https://github.com/coreos/fedora-coreos-config/pull/2785
so that we can get the full console from the first
boot on applehv.

For now, this is just a debugging aid for us.

Signed-off-by: Colin Walters <walters@verbum.org>