Remove --latest from podman CMD --help output
Because --latest is not supported on podman-remote commands
we should not be showing examples using podman-remote CMD --help
with --latest usage, it confuses users. Rather then hacking up
the code with if remote else --latest, it is better to just remove
information in help messages.

Prevents: https://github.com/containers/podman/issues/21174

[NO NEW TESTS NEEDED] Since normal tests should cover this.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>