Manually patch vendor/ to address CVE-2024-21626
We can't bump runc in this branch, it brings in too many other
things that won't build on this old Golang (this branch is still
on v1.16).

So, instead, backport the runc patches to this older runc.

The good news: We vendor very little of runc, so this was not all
that difficult.

Signed-off-by: Matt Heon <mheon@redhat.com>