Wave secrets protections on select files/scripts
The baseline gitleaks data is unfortunately commit-locked, meaning small
changes to files due to (for example) rebases, can render them useless.
Manually go through all findings and where possible mark lines to be
ignored directly.  In a few cases where secrets are used in tests, mark
them to be ignored via a new `.gitleaksignore` file.  This will
hopefully cut way down on the number of false-positive alerts that
require review.

Note: I intentionally did not wave checks in the `.cirrus.yml` file as
it's currently going through a large number of changes.  I'll leave it
up to a future followup commit to mark known/approved secret references
in this file.

[NO NEW TESTS NEEDED]

Signed-off-by: Chris Evich <cevich@redhat.com>